build: bump argo-cd to v3.4.6 for parity with the live server
v3.4 moved gitops-engine into the argo-cd repo as a nested module declared
with a local-path replace, which Go ignores for consumers and which carries
no tags, so it needs a consumer-side replace pinned to the v3.4.6 commit.

argo-cd v3.4.6 also forces gitlab-org/api/client-go to v1.x, whose IDs moved
from int to int64. The gitlab backend is fixed for compilation only - we run
GitHub, and none of these paths are exercised or tested here.

+++ pkg/vcs/gitlab_client/merge.go
@@ -36,7 +36,7 @@ func (c *Client) GetMergeChanges(ctx context.Context, projectId int, mergeReqId 
 	}
 
 	for {
-		diffs, resp, err := c.c.MergeRequests.ListMergeRequestDiffs(projectId, mergeReqId, opts)
+		diffs, resp, err := c.c.MergeRequests.ListMergeRequestDiffs(projectId, int64(mergeReqId), opts)
 		if err != nil {
 			telemetry.SetError(span, err, "Get MergeRequest Changes")
 			return changes, err
@@ -65,10 +65,10 @@ func (c *Client) GetMergeChanges(ctx context.Context, projectId int, mergeReqId 
 }
 
 type MergeRequestsServices interface {
-	GetMergeRequestDiffVersions(pid interface{}, mergeRequest int, opt *gitlab.GetMergeRequestDiffVersionsOptions, options ...gitlab.RequestOptionFunc) ([]*gitlab.MergeRequestDiffVersion, *gitlab.Response, error)
-	ListMergeRequestDiffs(pid interface{}, mergeRequest int, opt *gitlab.ListMergeRequestDiffsOptions, options ...gitlab.RequestOptionFunc) ([]*gitlab.MergeRequestDiff, *gitlab.Response, error)
-	UpdateMergeRequest(pid interface{}, mergeRequest int, opt *gitlab.UpdateMergeRequestOptions, options ...gitlab.RequestOptionFunc) (*gitlab.MergeRequest, *gitlab.Response, error)
-	GetMergeRequest(pid interface{}, mergeRequest int, opt *gitlab.GetMergeRequestsOptions, options ...gitlab.RequestOptionFunc) (*gitlab.MergeRequest, *gitlab.Response, error)
+	GetMergeRequestDiffVersions(pid interface{}, mergeRequest int64, opt *gitlab.GetMergeRequestDiffVersionsOptions, options ...gitlab.RequestOptionFunc) ([]*gitlab.MergeRequestDiffVersion, *gitlab.Response, error)
+	ListMergeRequestDiffs(pid interface{}, mergeRequest int64, opt *gitlab.ListMergeRequestDiffsOptions, options ...gitlab.RequestOptionFunc) ([]*gitlab.MergeRequestDiff, *gitlab.Response, error)
+	UpdateMergeRequest(pid interface{}, mergeRequest int64, opt *gitlab.UpdateMergeRequestOptions, options ...gitlab.RequestOptionFunc) (*gitlab.MergeRequest, *gitlab.Response, error)
+	GetMergeRequest(pid interface{}, mergeRequest int64, opt *gitlab.GetMergeRequestsOptions, options ...gitlab.RequestOptionFunc) (*gitlab.MergeRequest, *gitlab.Response, error)
 }
 
 type MergeRequestsService struct {
